feat: Adicionando as classes DTO´s do Autor, inserção da QueryMethod na interface LivroRepository visando evitar a violação de integridade referencial e ExceptionBaseDeletion com o status 409(conflict)

+++ src/main/java/com/devlib/devlib/services/AutorService.java
@@ -1,15 +1,17 @@
 package com.devlib.devlib.services;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.devlib.devlib.DTO.AutorInsertDTO;
+import com.devlib.devlib.DTO.AutorUpdateDTO;
 import com.devlib.devlib.entites.Autor;
 import com.devlib.devlib.repositories.AutorRepository;
+import com.devlib.devlib.repositories.LivroRepository;
+import com.devlib.devlib.services.exceptions.AutorDeletionException;
 import com.devlib.devlib.services.exceptions.AutorNotFoundException;
-import com.devlib.devlib.services.exceptions.BadRequestException;
 
 @Service
 public class AutorService {
@@ -17,40 +19,36 @@ public class AutorService {
 	@Autowired
 	private AutorRepository repository;
 	
+	@Autowired
+	private LivroRepository livroRepository;
+	
 	public List<Autor> findAll() {
 		return repository.findAll();
 	}
 	public Autor findById(Long id) {
-		validationAutorId(id);
-		Optional<Autor> autor = repository.findById(id);
-		return autor.get();
+		Autor autor = repository.findById(id)
+				.orElseThrow(() -> new AutorNotFoundException(id));
+		return autor;
 	}
-	public Autor insert(Autor autor) {
-		badRequestId(autor);
+	public Autor insert(AutorInsertDTO autorDTO) {
+		Autor autor = new Autor();
+		autor.setNome(autorDTO.getNome());
+		autor.setNacionalidade(autorDTO.getNacionalidade());
 	    return repository.save(autor);
 	}
-	public Autor update(Long id, Autor entity) {
-		validationAutorId(id);
-		Autor autor = repository.getReferenceById(id);
+	public Autor update(Long id, AutorUpdateDTO entity) {
+		Autor autor = repository.findById(id)
+				.orElseThrow(() -> new AutorNotFoundException(id));
 		updateData(autor, entity);
 		return repository.save(autor);
 	}
-	public void updateData(Autor autor, Autor entity) {
+	public void updateData(Autor autor, AutorUpdateDTO entity) {
 		autor.setNome(entity.getNome());
 		autor.setNacionalidade(entity.getNacionalidade());
 	}
 	public void delete(Long id) {
-		validationAutorId(id);
+		if(!repository.existsById(id)) throw new AutorNotFoundException(id);
+		if(livroRepository.existsByAutoresId(id)) throw new AutorDeletionException(id);
 		repository.deleteById(id);
-	}
-	public void validationAutorId(Long id){
-		if(!repository.existsById(id)) {
-			throw new AutorNotFoundException(id);
-		}
-	}
-	public void badRequestId(Autor autor) {
-		if (autor.getId() != null) {
-	        throw new BadRequestException();
-	    }
 	}
 }
